fixed casting bug

+++ src/print.h
@@ -15,17 +15,17 @@
 
 #define PrintMemCols8(_b_, _s_, _a_) \
 { \
-    UINT64 _hw_v_ = (SIZE_T)_a_ + _s_; \
+    UINT64 _hw_v_ = (UINT64)_a_ + _s_; \
     UINT8 _hw_w_ = 0x10; \
     HEX_CHAR_WIDTH(_hw_v_, _hw_w_); \
     \
-    for ( ULONG _i_ = 0; _i_ < _s_; _i_+=0x10 ) \
+    for ( UINT64 _i_ = 0; _i_ < _s_; _i_+=0x10 ) \
     { \
-        ULONG _end_ = (_i_+0x10<_s_) ? (_i_+0x10) : (_s_); \
+        UINT64 _end_ = (_i_+0x10<_s_) ? (_i_+0x10) : (_s_); \
         ULONG _gap_ = (_i_+0x10<=_s_) ? 0 : ((0x10+_i_-_s_)*3); \
-        printf("%.*zx  ", _hw_w_, (((SIZE_T)_a_)+_i_)); \
+        printf("%.*zx  ", _hw_w_, (((UINT64)_a_)+_i_)); \
          \
-        for ( ULONG _j_ = _i_, _k_=0; _j_ < _end_; _j_++, _k_++ ) \
+        for ( UINT64 _j_ = _i_, _k_=0; _j_ < _end_; _j_++, _k_++ ) \
         { \
             printf("%02x", ((PUINT8)_b_)[_j_]); \
             printf("%c", (_k_==7?'-':' ')); \
@@ -35,7 +35,7 @@
             printf(" "); \
         } \
         printf("  "); \
-        for ( ULONG _k_ = _i_; _k_ < _end_; _k_++ ) \
+        for ( UINT64 _k_ = _i_; _k_ < _end_; _k_++ ) \
         { \
             if ( ((PUINT8)_b_)[_k_] < 0x20 || ((PUINT8)_b_)[_k_] > 0x7E || ((PUINT8)_b_)[_k_] == 0x25 ) \
             { \
@@ -52,17 +52,17 @@
 
 #define PrintMemCols16(_b_, _s_, _a_) \
 { \
-    UINT64 _hw_v_ = (SIZE_T)_a_ + _s_; \
+    UINT64 _hw_v_ = (UINT64)_a_ + _s_; \
     UINT8 _hw_w_ = 0x10; \
     HEX_CHAR_WIDTH(_hw_v_, _hw_w_); \
     \
-    for ( ULONG _i_ = 0; _i_ < _s_; _i_+=0x10 ) \
+    for ( UINT64 _i_ = 0; _i_ < _s_; _i_+=0x10 ) \
     { \
-        ULONG _end_ = (_i_+0x10<_s_)?(_i_+0x10):(_s_); \
+        UINT64 _end_ = (_i_+0x10<_s_)?(_i_+0x10):(_s_); \
         ULONG _gap_ = (_i_+0x10<=_s_) ? 0 : ((0x10+_i_-_s_)/2*5); \
-        printf("%.*zx  ", _hw_w_, (((SIZE_T)_a_)+_i_)); \
+        printf("%.*zx  ", _hw_w_, (((UINT64)_a_)+_i_)); \
          \
-        for ( ULONG _j_ = _i_; _j_ < _end_; _j_+=2 ) \
+        for ( UINT64 _j_ = _i_; _j_ < _end_; _j_+=2 ) \
         { \
             printf("%04x ", *(PUINT16)&(((PUINT8)_b_)[_j_])); \
         } \
@@ -71,7 +71,7 @@
             printf(" "); \
         } \
         printf("  "); \
-        for ( ULONG _j_ = _i_; _j_ < _end_; _j_+=2 ) \
+        for ( UINT64 _j_ = _i_; _j_ < _end_; _j_+=2 ) \
         { \
             printf("%wc", *(PUINT16)&(((PUINT8)_b_)[_j_])); \
         } \
@@ -81,16 +81,16 @@
 
 #define PrintMemCols32(_b_, _s_, _a_) \
 {\
-    UINT64 _hw_v_ = (SIZE_T)_a_ + _s_; \
+    UINT64 _hw_v_ = (UINT64)_a_ + _s_; \
     UINT8 _hw_w_ = 0x10; \
     HEX_CHAR_WIDTH(_hw_v_, _hw_w_); \
     \
-    for ( ULONG _i_ = 0; _i_ < _s_; _i_+=0x10 ) \
+    for ( UINT64 _i_ = 0; _i_ < _s_; _i_+=0x10 ) \
     { \
-        ULONG _end_ = (_i_+0x10<_s_)?(_i_+0x10):(_s_); \
-        printf("%.*zx  ", _hw_w_, (((SIZE_T)_a_)+_i_)); \
+        UINT64 _end_ = (_i_+0x10<_s_)?(_i_+0x10):(_s_); \
+        printf("%.*zx  ", _hw_w_, (((UINT64)_a_)+_i_)); \
          \
-        for ( ULONG _j_ = _i_; _j_ < _end_; _j_+=4 ) \
+        for ( UINT64 _j_ = _i_; _j_ < _end_; _j_+=4 ) \
         { \
             printf("%08x ", *(PUINT32)&(((PUINT8)_b_)[_j_])); \
         } \
@@ -100,16 +100,16 @@
 
 #define PrintMemCols64(_b_, _s_, _a_) \
 {\
-    UINT64 _hw_v_ = (SIZE_T)_a_ + _s_; \
+    UINT64 _hw_v_ = (UINT64)_a_ + _s_; \
     UINT8 _hw_w_ = 0x10; \
     HEX_CHAR_WIDTH(_hw_v_, _hw_w_); \
     \
-    for ( ULONG _i_ = 0; _i_ < _s_; _i_+=0x10 ) \
+    for ( UINT64 _i_ = 0; _i_ < _s_; _i_+=0x10 ) \
     { \
-        ULONG _end_ = (_i_+0x10<_s_)?(_i_+0x10):(_s_); \
-        printf("%.*zx  ", _hw_w_, (((SIZE_T)_a_)+_i_)); \
+        UINT64 _end_ = (_i_+0x10<_s_)?(_i_+0x10):(_s_); \
+        printf("%.*zx  ", _hw_w_, (((UINT64)_a_)+_i_)); \
          \
-        for ( ULONG _j_ = _i_; _j_ < _end_; _j_+=8 ) \
+        for ( UINT64 _j_ = _i_; _j_ < _end_; _j_+=8 ) \
         { \
             printf("%016llx ", *(PUINT64)&(((PUINT8)_b_)[_j_])); \
         } \
@@ -119,7 +119,7 @@
 
 #define PrintMemBytes(_b_, _s_) \
 { \
-    for ( ULONG _i_ = 0; _i_ < _s_; _i_++ ) \
+    for ( UINT64 _i_ = 0; _i_ < _s_; _i_++ ) \
     { \
         printf("%02x ", ((PUINT8)_b_)[_i_]); \
     } \
